Adds router, refactors table view
