+++ app/src/Pages/Portal/Expenses/Expenses.js
@@ -1,18 +1,37 @@
 import * as React from "react";
-import PageHoc from "../../../components/PageHoc/PageHoc";
-// import AddExpense from "./AddExpense/AddExpense";
-// import ExpensesDetail from "./ExpensesDetail/ExpensesDetail";
-import ExpensesTable from "./ExpensesTable/ExpensesTable";
+import { Route, Switch } from "react-router-dom";
+//import PageHoc from "../../../components/PageHoc/PageHoc";
+import AddExpense from "./AddExpense/AddExpense";
+import ExpensesDetail from "./ExpensesDetail/ExpensesDetail";
+import ExpensesTable from "./ExpensesTable";
 
 // Switch statement for routing
 
 const Expense = props => {
+  const { match } = props
   return (
-    <PageHoc>
-      {/* <AddExpense /> */}
-      {/* <ExpensesDetail /> */}
-      <ExpensesTable  {...props} />
-    </PageHoc>
+    <Route 
+      render={({ location }) => {
+        return (
+        <Switch location={location}>
+          <Route 
+            exact 
+            path={`${match.url}/add`}
+            component={AddExpense}
+          />
+          <Route 
+            exact 
+            path={`${match.url}/:id`}
+            component={ExpensesDetail}
+          />
+          <Route 
+            exact 
+            path={`${match.url}/`}
+            component={ExpensesTable} 
+          />
+        </Switch>
+      )}}
+    />
   );
 };
 export default Expense;
